Implementing handling of assymetric energies
The beam energies must be specified in the slight.in file as
BEAM_1_GAMMA and BEAM_2_GAMMA, even if the energies are symmetric.

The beam lorentz gamma for the CMS system is calculated at
initialization, and every calculation is based on that specific
lorentz gamma. At the end of the event generation the output is
boosted to the lab system. This should guarantee that no changes
are required to the rest of the code, which can be used as is.