Moving calculation of lookup table to seperate method.
This is done to be able to make probabilityOfBreakup(...)
a const method. The number of bins used in the lookup table
is now determined by a cut off in accuracy instead of a fixed
number of bins.

There is also an additional calculation of breakup probabilities
for p-A collisions implemented.