Added the pickw() and picky() into the do-while loops.
Initially I never knew imagined that the pickw and picky were not in the gammasingle while loops
The insight came after testing completely  ZOVERZ03, while try to test F2 and F2Prime section.
The computer kept stuck inthe do while loop, I changed Pid to 335 thinking it was the pid. Nothing changed
I then removed eta_cuts and observed the plot of pseudorapidity and found it to be abundant in the needed range
I debugged line by line. with cuts there was truly no fulfilling particle, but without cuts there were several particles that both satisfied constraints.
Thats when I sensed that, and got insight that it was pickw and picky.
After fixing it It couldnt work again. It got stuck. The mistake was that I mistakenly set CUT_PT instead of CUT_ETA.
Before I discovered it this time it was a tug of war. I debugged and while I corrected and moved the initial pickw and picky to the pythia_decay section
To ensure that both CUT=0 and CUT =1 cases spill the same particles.
Thats when God released insight that I made that mistake. I fixed this Pythia_decay issue and the CUT_PT mistake also.
And I have tested so far and it worked.

	modified:   ../src/gammagammasingle.cpp
	../config/slight.in.rho_rhic_cuteta
	../config/slightb4.in