Updated HEPMC3 implementation.
HEPMC cannot be implemented in the ASCii output but extra beam information can be provided. Functionality for extra beam info is now included.
HEPMC3 parameter is no longer settable in the slight.in. It is now a full blown -runtime argument/parameter
It is still left as a parameter, because in the case of interference for example, HEPMC3 can be step but unimplementable.
giveExtrabeamInfo global variable is also created to track if writing extra beam infor is possible.
The importance of the association between the gamma's and the emmiting beam (especially in Two Photon interactions) has been identified and tracked in the order of with which beams and gammas are added to their respective list and written out in the output event.
Documentation for important functions in the upcXevent has been improved for clarity.
The wrong association between the squared transfered momentum "t" and the target beams which I assumed from wrong interpretation of estarlight object design has been removed. As it prevents having two sources, and zero targets (and if there are two targets, there must be two "t"s - which is wrong)

 Changes to be committed:
	modified:   config/slight.in
	modified:   config/slight.in.Fig_1
	modified:   config/slight.in.Fig_4
	modified:   config/slight.in.Fig_5
	modified:   config/slight.in.ee_rhic
	modified:   config/slight.in.jpsi_lhc
	modified:   config/slight.in.pPb_lhc
	modified:   config/slight.in.rho_rhic
	modified:   include/inputParameters.h
	modified:   include/upcXevent.h
	modified:   src/eventfilewriter.cpp
	modified:   src/gammaavm.cpp
	modified:   src/gammagammaleptonpair.cpp
	modified:   src/gammagammasingle.cpp
	modified:   src/inputParameters.cpp